requirements.txt
fastapi==0.135.2
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
uvicorn[standard]==0.42.0
psycopg2-binary==2.9.11 
sqlalchemy==2.0.49
python-jose==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.24